+++ babel.config.js
@@ -7,9 +7,15 @@ const moduleSystem = (BABEL_ENV && BABEL_ENV.startsWith('modules:')) ? BABEL_ENV
 
 // For the ES configuration only transpile react to valid JavaScript.
 // For commonjs transpile to old JS versions.
-const presets = moduleSystem === "es"
-  ? ['@babel/preset-react']
-  : [
+const presets  = [
+  ['@babel/preset-typescript', {
+    allExtensions: true,
+    isTSX: true
+  }],
+  '@babel/preset-react'
+];
+if (moduleSystem !== "es") {
+  presets.splice(0, 0,
     ['@babel/preset-env', {
       targets: {
         ie: 11,
@@ -20,10 +26,15 @@ const presets = moduleSystem === "es"
         node: '6.11',
       },
       modules: moduleSystem,
-    }],
-    '@babel/preset-react'
-  ];
+    }]
+  );
+}
 
+// The ES system does not polyfill etc, while the others do.
+const transformOptions = {
+  helpers: true,
+  useESModules: moduleSystem === "es"
+};
 
 module.exports = {
   presets,
@@ -31,19 +42,16 @@ module.exports = {
     '@babel/plugin-transform-object-assign',
     '@babel/plugin-proposal-object-rest-spread',
     'babel-plugin-extensible-destructuring',
-    [
-      // The ES system does not polyfill etc, while the others do
-      '@babel/plugin-transform-runtime',
-      {
-        helpers: true,
-        useESModules: moduleSystem === 'es'
-      }
-    ],
+
+    // Stage 0
+    ['@babel/plugin-transform-runtime', transformOptions],
+
+    // Stage 1
+
+    // Stage 2
+
     // Stage 3
-    [
-      '@babel/plugin-proposal-class-properties',
-      { loose: true }
-    ],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
   ],
   env: {
     production: {