+++ src/err.ts
@@ -1,4 +1,4 @@
-import { getType, locate } from 'mitsuketa';
+import { getType, locate } from './mitsuketa';
 
 const UNKNOWN = '<unknown parameter>';
 
@@ -30,74 +30,73 @@ export function isFalsy(paramName = UNKNOWN, param: any) {
   }
 }
 
-export function isNoneOf(paramName: string = UNKNOWN, param: any, contains: Array<any> = []) {
+export function isNoneOf(paramName = UNKNOWN, param: any, contains: Array<any> = []) {
   if (!contains.includes(param)) {
     throwError(getCaller(2), paramName, `to be any of ${JSON.stringify(contains)}`);
   }
 }
 
-export function isAnyOf(paramName: string = UNKNOWN, param: any, contains: Array<any> = []) {
+export function isAnyOf(paramName = UNKNOWN, param: any, contains: Array<any> = []) {
   if (contains.includes(param)) {
     throwError(getCaller(2), paramName, `not to be any of ${JSON.stringify(contains)}`);
   }
 }
 
-export function isNotType(paramName: string = UNKNOWN, param: any, type = '') {
+export function isNotType(paramName = UNKNOWN, param: any, type = '') {
   if (getType(param) !== type.toLowerCase()) {
     throwError(getCaller(2), paramName, `to be type ${type.toLowerCase()}`);
   }
 }
 
-export function isAnyTypeOf(paramName: string = UNKNOWN, param: any, types: Array<string> = []) {
+export function isAnyTypeOf(paramName = UNKNOWN, param: any, types: Array<string> = []) {
   types.forEach(type => {
-      if (getType(param) === type) {
-        throwError(getCaller(2), paramName, `not to be type of ${type.toLowerCase()}`);
-      }
+    if (getType(param) === type) {
+      throwError(getCaller(2), paramName, `not to be type of ${type.toLowerCase()}`);
+    }
   });
 }
 
-export function missingKey(paramName: string = UNKNOWN, param: any, keyName = '') {
+export function missingKey(paramName = UNKNOWN, param: any, keyName = '') {
   isUndefined(paramName, param);
   if (!Object.keys(param).includes(keyName)) {
     throwError(getCaller(2), paramName, `to contain '${keyName}' key`);
   }
 }
 
-export function missingAnyKeys(paramName: string = UNKNOWN, param: any, keyNames: Array<string> = ['']) {
+export function missingAnyKeys(paramName = UNKNOWN, param: any, keyNames: Array<string> = []) {
   isUndefined(paramName, param);
   const keyList = Object.keys(param);
   keyNames.forEach(keyName => {
-      if (!keyList.includes(keyName)) {
-        throwError(getCaller(2), paramName, `to contain '${keyName}' key`);
-      }
+    if (!keyList.includes(keyName)) {
+      throwError(getCaller(2), paramName, `to contain '${keyName}' key`);
+    }
   });
 }
 
-export function containsUndefined(paramName: string = UNKNOWN, param: any) {
+export function containsUndefined(paramName = UNKNOWN, param: any) {
   [undefined, null].forEach(value => {
-      const location = locate(param, value, null);
-      if (location) {
-        throwError(getCaller(2), paramName, `not to contain '${JSON.stringify(value)}' at ${location}`);
-      }
+    const location = locate(param, value);
+    if (location) {
+      throwError(getCaller(2), paramName, `not to contain '${JSON.stringify(value)}' at ${location}`);
+    }
   });
 }
 
-export function isInvalidPath(paramName: string = UNKNOWN, param: any) {
+export function isInvalidPath(paramName = UNKNOWN, param: any) {
   isUndefined(paramName, param);
   isNotType(paramName, param, 'string');
   isAnyOf(paramName, param, ['', '/']);
   ['.', '$', '[', ']', '#'].forEach(invalidChar => {
-      if (param.indexOf(invalidChar) > -1) {
-        throwError(getCaller(2), paramName, `not to contain invalid character '${invalidChar}'`);
-      }
+    if (param.indexOf(invalidChar) > -1) {
+      throwError(getCaller(2), paramName, `not to contain invalid character '${invalidChar}'`);
+    }
   });
-
   if (param.match(/\/{2,}/g)) {
     throwError(getCaller(2), paramName, 'not to contain consecutive forward slash characters');
   }
 }
 
-export  function isInvalidWriteData(paramName: string = UNKNOWN, param: any) {
+export function isInvalidWriteData(paramName = UNKNOWN, param: any) {
   isUndefined(paramName, param);
   containsUndefined(paramName, param);
 }