+++ src/locale/es.ts
@@ -1,4 +1,4 @@
-import Locale from './interface';
+import { Locale } from '.';
 
 // Spanish
 const locale: Locale = {