+++ src/locale/ru.ts
@@ -1,4 +1,4 @@
-import Locale from './interface';
+import { Locale } from '.';
 
 // Russian
 const locale: Locale = {